LPD-60003 move remainder files to product-analytics folder
