+++ modules/test/playwright/pages/product-analytics-web/ProductAnalyticsBannerPage.ts
