+++ modules/test/playwright/pages/product-analytics-web/ProductAnalyticsConsentPanelPage.ts
